+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- setting to use hardware next/previous buttons to skip chapters or rewind/fast forward
+
+### Fixed
+
+- downloaded items failing to play after path migration
+
 ## [v0.10.1] - 2026-09-19
 
 ### Added